requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
requests==2.31.0
SQLAlchemy==2.0.19
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
